+++ src/pages/pageIndex.ts
@@ -2,6 +2,6 @@ import Home from "./Home";
 import LoginPage from "./Login";
 import NewEvaluation from "./NewEvaluation";
 import AuditeeDashboard from "./AuditeeDashboard";
-import NewAuditee from "./NewAuditee";
+import AuditeeForm from "./AuditeeForm";
 
-export { Home, LoginPage, NewEvaluation, AuditeeDashboard, NewAuditee };
+export { Home, LoginPage, NewEvaluation, AuditeeDashboard, AuditeeForm };
