+++ src/services/companyServices.ts
@@ -23,6 +23,24 @@ export class CompanyService {
 		}
 	}
 
+	async getCompanyByCompanyId(tenantId: string, companyId: string): Promise<CompanyListDto | any> {
+		try {
+			const response = await axios.get<CompanyListDto>(
+				`https://ryzr-be-cwacd8a5c6c8d7bd.francecentral-01.azurewebsites.net/api/v1/companies/${tenantId}/${companyId}`
+			);
+			if (response.status !== 200) {
+				throw response;
+			}
+			return response.data;
+		} catch (error) {
+			const errorInfo = handleAxiosError(error);
+			console.error("Error fetching company:", errorInfo.message);
+
+			//rethrowing for conditional rendering
+			throw errorInfo;
+		}
+	}
+
 	async createCompany(
 		companyData: CompanyCreateDto
 	): Promise<CompanyListDto | any> {
